+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.10] - 2026-01-25
+
+### Fixed
+- `TableBody` not showing filler row cells if `columnOrder` does not exist
+
 ## [0.6.9] - 2026-01-25
 
 ### Added